requirements.txt
addict==2.3.0
alabaster==0.7.12
asciitree==0.3.3
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
configparser==5.0.1
coverage==5.3
cssselect2==0.4.1
# -e git+https://github.com/mila-iqia/COVI-ML@19986f7427a7a643eb05fb41e5ed4dd113362cd6#egg=ctt
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
docker-pycreds==0.4.0
docutils==0.16
fasteners==0.15
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
h5py==3.1.0
html5lib==1.1
hydra-core==0.11.0
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
monotonic==1.5
networkx==2.4
numcodecs==0.7.2
numpy==1.19.4
omegaconf==1.4.1
onnx==1.8.0
onnx-tf @ git+git://github.com/onnx/onnx-tensorflow.git@c4fc047e7ec71daa6aa8f71e9cc2ee9e5a3768b6
orderedset==2.0.3
packaging==20.4
pandas==1.1.4
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
Pyphen==0.10.0
pytest==6.1.2
pytest-cov==2.8.1
pytest-subtests==0.3.1
pytest-timeout==1.4.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
recommonmark==0.6.0
requests==2.25.0
runstats==1.8.0
scipy==1.5.4
seaborn==0.11.0
sentry-sdk==0.19.3
shortuuid==1.0.1
simpy==4.0.1
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
speedrun @ git+git://github.com/inferno-pytorch/speedrun.git@3570bf47196fb9a1406f600d055160aec6a71e3c
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subprocess32==3.5.4
tinycss2==1.1.0
toml==0.10.2
torch==1.7.0
tqdm==4.52.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wandb==0.10.11
watchdog==0.10.3
wcwidth==0.2.5
WeasyPrint==52.1
webencodings==0.5.1
xdelta3==0.0.5
zarr==2.5.0
zipfile36==0.1.3
Flask==2.0.2
